# api/swagger.yaml
...
swagger: "2.0"
schemes:
  - "http"
  - "https"
produces:
  - "application/json"
consumes:
  - "application/json"
  - "application/x-www-form-urlencoded"
basePath: "/oauth"
info:
  title: "OAuth API"
  version: "1.0.0"
  description: |
    The OAuth API provides OAuth Bearer Token management as well as
    session and cookie mangement for browser based flows.

    # Errors

    The API uses standard HTTP status codes to indicate the success or failure
    of the API call. The body of the response will be JSON in the following
    format:
    ```
    {
      "message": "object not found"
    }
    ```

    # Reference Definitions

      This API provides reference definitions to be used by go projects to extend 
      the base types to suit the needs of the backend. For example:

      definitions:
        Application:
          description: |
            Applications are clients for the atomic API
          allOf:
            - $ref: '/go/src/github.com/libatomic/oauth/api/swagger.yaml#/definitions/ApplicationRef'
            - type: 'object'
              properties:
                id:
                  type: 'string'
                  description: 'The application id'
                created_at:
                  type: 'string'
                  description: 'The application creation date'
                  format: 'date-time'
                updated_at:
                  type: 'string'
                  description: 'The last update date'
...

...
      accessible.
...
securityDefinitions:
  OAuth:
    type: oauth2
    flow: accessCode
    authorizationUrl: https://auth.server.local/oauth2/authorize
    tokenUrl: https://auth.server.local/oauth2/token
    scopes:
      openid: Access a user's identity
      profile: Access a user's profile
      offline_access: Used by clients to request permission to use refresh tokens
definitions:
...
  ErrorResponse:
    description: "A common error response"
    type: "object"
    required: ["message"]
    properties:
      message:
        type: "string"
        description: "The error message"
        x-nullable: false
paths:
...
  /authorize:
    get:
      description: |
        Gererates an oauth authorization request for login and redirects the browser to the login form.

        This method will redirect the browser to `app_uri` with an encoded `request_token` parameter, which will present the user with a form.

          `app_uri` and `redirect_uri` must be registered with the api client.

        The form could post back to `/login` or `/signup` which will perform the final redirect to `redirect_api` with the authorization code.

        Errors will be delivered to `redirect_uri` with two query parameters:


          - `error=<error name>` (i.e. `bad_request`, etc) 
          - `error_description=<error detail>`
      operationId: "Authorize"
      tags: ["Auth"]
      parameters:
        - name: "response_type"
          in: "query"
          description: "The authorization code response type"
          type: "string"
          enum: ["code"]
          required: true
        - name: "client_id"
          in: "query"
          description: "The client id"
          type: "string"
          required: true
        - name: "audience"
          in: "query"
          type: "string"
          required: true
        - name: "user_pool"
          in: query
          description: "The user pool to authorize against"
          type: "string"
          required: false
        - name: "app_uri"
          in: query
          description: "The URL to which the authentication server redirects the browser for action"
          type: "string"
          required: false
        - name: "redirect_uri"
          in: "query"
          description: "The URL to which the authentication server redirects the browser after authorization has been granted by the user"
          type: "string"
          required: false
        - name: "scope"
          in: "query"
          description: |
            The requested scopes, if empty will request all the user permissions.
          type: "array"
          items:
            type: "string"
          collectionFormat: "ssv"
        - name: "state"
          in: "query"
          description: "Opaque state returned the redirect uri"
          type: "string"
        - name: "code_challenge_method"
          in: "query"
          description: |
            The method used to generate the challenge. The PKCE RFC defines two methods, S256 and plain; 
            however, the authentication serivce supports only S256.
          type: "string"
          enum: ["S256"]
          default: "S256"
        - name: "code_challenge"
          in: "query"
          description: "The generated challenge from the code_verifier."
          type: "string"
          required: true
      responses:
        302:
          description: "found"
          headers:
            Location:
              type: "string"
              description: "The location redirect header"
        400:
          description: "bad parameter"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "unauthorized"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "internal server error"
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /.well-known/jwks.json:
    get:
      description: |
        Return the jwks which includes the rsa public key and ids for the service
      operationId: "PublicKeyGet"
      tags: ["Auth"]
      responses:
        200:
          description: "ok"
          schema:
            type: "string"
        500:
          description: "internal server error"
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /userInfo:
    get:
      description: |
        Get the current user profile data
      operationId: "UserInfoGet"
      tags: ["User"]
      responses:
        200:
          description: "ok"
          schema:
            $ref: "#/definitions/Profile"
        400:
          description: "bad parameter"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "unauthorized"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "internal server error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
        - OAuth: ["openid", "profile"]
    put:
      description: |
        Update the user's info
      operationId: "UserInfoUpdate"
      tags: ["User"]
      parameters:
        - name: "profile"
          description: "The new profile"
          in: "body"
          schema:
            $ref: "#/definitions/Profile"
      responses:
        200:
          description: "ok"
          schema:
            $ref: "#/definitions/Profile"
        400:
          description: "bad parameter"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "unauthorized"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "internal server error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
        - OAuth: ["openid", "profile"]